upgrades/requirements_upgrade.txt
# StealthMaster Upgrade Requirements
# ==================================

# Core dependencies
playwright>=1.40.0
selenium>=4.15.0
undetected-chromedriver>=3.5.4
nodriver>=0.25  # Advanced undetected browser

# CloudFlare bypass
docker>=6.1.3  # For FlareSolverr container
cloudscraper>=1.2.71  # Backup CloudFlare bypass

# Performance and async
aiohttp>=3.9.1
asyncio>=3.4.3
uvloop>=0.19.0  # Faster event loop
aiofiles>=23.2.1

# Browser automation enhancements
playwright-stealth>=1.0.6
selenium-stealth>=1.0.6
pyppeteer>=1.0.2
browserforge>=1.0.0  # Browser fingerprinting

# Proxy management
aiohttp-socks>=0.8.4
python-socks>=2.4.4
proxy-randomizer>=1.3.0

# Resource monitoring
psutil>=5.9.6
memory-profiler>=0.61.0
py-spy>=0.3.14  # Performance profiling

# Data handling
beautifulsoup4>=4.12.2
lxml>=4.9.3
selectolax>=0.3.17  # Fast HTML parsing

# Italian-specific
italian-fiscal-code>=0.2.1  # For Italian tax code validation

# WebSocket support
websockets>=12.0
python-socketio>=5.10.0

# Captcha solving
2captcha-python>=1.2.5
anticaptchaofficial>=1.0.53

# Enhanced logging
loguru>=0.7.2
python-json-logger>=2.0.7

# Testing and debugging
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-playwright>=0.4.3

# Monitoring and metrics
prometheus-client>=0.19.0
grafana-api>=1.0.3

# Additional utilities
python-dotenv>=1.0.0
pydantic>=2.5.2
tenacity>=8.2.3  # Retry logic
schedule>=1.2.0
humanize>=4.9.0

# Image processing (for captchas)
Pillow>=10.1.0
opencv-python>=4.8.1.78

# Machine learning (for behavior patterns)
scikit-learn>=1.3.2
numpy>=1.26.2
pandas>=2.1.4

# Encryption and security
cryptography>=41.0.7
pycryptodome>=3.19.0

# Development tools
black>=23.12.0
pylint>=3.0.3
mypy>=1.7.1
pre-commit>=3.6.0
